World Save Issues Step 7 - Update EventListener to utilize previous changes and split effects that only work on `Living` entities for PR https://github.com/DaPigGuy/PiggyCustomEnchants/pull/122
A lot happens in this update. First, the check for `$entity instanceof Living` in `checkGlobalEnchants()` is split to its own check instead of combined with `$event instanceof EntityDamageEvent` because there are some effects that do not require '$entity' to be living.  After that check is split, the effect are split appropriately where any effect that requires the use of `addEffect()`,`getEffect()`, or `hasEffect()` are only run `if($entity instanceof Living)`.  Then all other effect are executed.

Second major change here is the return to using `Entity::createEntity` inplace of just creating new objects.  This fixes a problem where when saving a level and PiggyProjectiles are still present, the server crashes because the projectiles aren't registered.